store and delete county or city

+++ laravel-zip-api/database/migrations/2025_09_29_074502_create_cities_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('postal_code');
             $table->string('name');
-            $table->integer('county_id');
+            $table->bigInteger('county_id');
 
             $table->foreign('county_id')->references('id')->on('counties')->onDelete('restrict');
         });
